Week10/Day1/requirements.txt
argparse
deepeval
fastapi
langfuse
openai
openai-agents
python-dotenv
uvicorn
gradio
pydantic
requests
